Removed default impl for build_empty_payload
