+++ crates/payload/basic/src/lib.rs
@@ -19,25 +19,18 @@ use reth_payload_builder::{
 };
 use reth_primitives::{
     bytes::BytesMut,
-    constants::{
-        BEACON_NONCE, EMPTY_RECEIPTS, EMPTY_TRANSACTIONS, EMPTY_WITHDRAWALS,
-        ETHEREUM_BLOCK_GAS_LIMIT, RETH_CLIENT_VERSION, SLOT_DURATION,
-    },
-    proofs, Block, BlockNumberOrTag, Bytes, ChainSpec, Header, Receipts, SealedBlock, Withdrawal,
-    B256, EMPTY_OMMER_ROOT_HASH, U256,
+    constants::{EMPTY_WITHDRAWALS, ETHEREUM_BLOCK_GAS_LIMIT, RETH_CLIENT_VERSION, SLOT_DURATION},
+    proofs, BlockNumberOrTag, Bytes, ChainSpec, SealedBlock, Withdrawal, B256, U256,
 };
 use reth_provider::{
-    BlockReaderIdExt, BlockSource, BundleStateWithReceipts, CanonStateNotification, ProviderError,
-    StateProviderFactory,
+    BlockReaderIdExt, BlockSource, CanonStateNotification, ProviderError, StateProviderFactory,
 };
-use reth_revm::{
-    database::StateProviderDatabase,
-    state_change::{apply_beacon_root_contract_call, post_block_withdrawals_balance_increments},
+use reth_revm::state_change::{
+    apply_beacon_root_contract_call, post_block_withdrawals_balance_increments,
 };
 use reth_tasks::TaskSpawner;
 use reth_transaction_pool::TransactionPool;
 use revm::{
-    db::states::bundle_state::BundleRetention,
     primitives::{BlockEnv, CfgEnv, Env},
     Database, DatabaseCommit, State,
 };
@@ -152,7 +145,7 @@ where
     Client: StateProviderFactory + BlockReaderIdExt + Clone + Unpin + 'static,
     Pool: TransactionPool + Unpin + 'static,
     Tasks: TaskSpawner + Clone + Unpin + 'static,
-    Builder: PayloadBuilder<Pool, Client> + Unpin + 'static,
+    Builder: PayloadBuilder<Pool, Client, BuiltPayload = EthBuiltPayload> + Unpin + 'static,
     <Builder as PayloadBuilder<Pool, Client>>::Attributes: Unpin + Clone,
     <Builder as PayloadBuilder<Pool, Client>>::BuiltPayload: Unpin + Clone,
 {
@@ -458,7 +451,7 @@ where
     Client: StateProviderFactory + Clone + Unpin + 'static,
     Pool: TransactionPool + Unpin + 'static,
     Tasks: TaskSpawner + Clone + 'static,
-    Builder: PayloadBuilder<Pool, Client> + Unpin + 'static,
+    Builder: PayloadBuilder<Pool, Client, BuiltPayload = EthBuiltPayload> + Unpin + 'static,
     <Builder as PayloadBuilder<Pool, Client>>::Attributes: Unpin + Clone,
     <Builder as PayloadBuilder<Pool, Client>>::BuiltPayload: Unpin + Clone,
 {
@@ -789,100 +782,14 @@ pub trait PayloadBuilder<Pool, Client>: Send + Sync + Clone {
         None
     }
 
-    /// Builds an empty payload without any transactions.
-    fn build_empty_payload<Attributes, Payload>(
+    /// Builds an empty payload without any transaction.
+    fn build_empty_payload<Attributes>(
         client: &Client,
         config: PayloadConfig<Attributes>,
-    ) -> Result<Payload, PayloadBuilderError>
+    ) -> Result<EthBuiltPayload, PayloadBuilderError>
     where
         Client: StateProviderFactory,
-        Attributes: PayloadBuilderAttributes,
-        Payload: BuiltPayload,
-    {
-        let extra_data = config.extra_data();
-        let PayloadConfig {
-            initialized_block_env,
-            parent_block,
-            attributes,
-            chain_spec,
-            initialized_cfg,
-            ..
-        } = config;
-
-        debug!(target: "payload_builder", parent_hash = ?parent_block.hash, parent_number = parent_block.number, "building empty payload");
-
-        let state = client.state_by_block_hash(parent_block.hash).map_err(|err| {
-            warn!(target: "payload_builder", parent_hash=%parent_block.hash, ?err,  "failed to get state for empty payload");
-            err
-        })?;
-        let mut db = State::builder()
-            .with_database_boxed(Box::new(StateProviderDatabase::new(&state)))
-            .with_bundle_update()
-            .build();
-
-        let base_fee = initialized_block_env.basefee.to::<u64>();
-        let block_number = initialized_block_env.number.to::<u64>();
-        let block_gas_limit: u64 = initialized_block_env.gas_limit.try_into().unwrap_or(u64::MAX);
-
-        // apply eip-4788 pre block contract call
-        pre_block_beacon_root_contract_call(
-            &mut db,
-            &chain_spec,
-            block_number,
-            &initialized_cfg,
-            &initialized_block_env,
-            &attributes,
-        ).map_err(|err| {
-            warn!(target: "payload_builder", parent_hash=%parent_block.hash, ?err,  "failed to apply beacon root contract call for empty payload");
-            err
-        })?;
-
-        let WithdrawalsOutcome { withdrawals_root, withdrawals } =
-            commit_withdrawals(&mut db, &chain_spec, attributes.timestamp(), attributes.withdrawals().clone()).map_err(|err| {
-                warn!(target: "payload_builder", parent_hash=%parent_block.hash,?err,  "failed to commit withdrawals for empty payload");
-                err
-            })?;
-
-        // merge all transitions into bundle state, this would apply the withdrawal balance changes
-        // and 4788 contract call
-        db.merge_transitions(BundleRetention::PlainState);
-
-        // calculate the state root
-        let bundle_state =
-            BundleStateWithReceipts::new(db.take_bundle(), Receipts::new(), block_number);
-        let state_root = state.state_root(&bundle_state).map_err(|err| {
-            warn!(target: "payload_builder", parent_hash=%parent_block.hash, ?err,  "failed to calculate state root for empty payload");
-            err
-        })?;
-
-        let header = Header {
-            parent_hash: parent_block.hash,
-            ommers_hash: EMPTY_OMMER_ROOT_HASH,
-            beneficiary: initialized_block_env.coinbase,
-            state_root,
-            transactions_root: EMPTY_TRANSACTIONS,
-            withdrawals_root,
-            receipts_root: EMPTY_RECEIPTS,
-            logs_bloom: Default::default(),
-            timestamp: attributes.timestamp(),
-            mix_hash: attributes.prev_randao(),
-            nonce: BEACON_NONCE,
-            base_fee_per_gas: Some(base_fee),
-            number: parent_block.number + 1,
-            gas_limit: block_gas_limit,
-            difficulty: U256::ZERO,
-            gas_used: 0,
-            extra_data,
-            blob_gas_used: None,
-            excess_blob_gas: None,
-            parent_beacon_block_root: attributes.parent_beacon_block_root(),
-        };
-
-        let block = Block { header, body: vec![], ommers: vec![], withdrawals };
-        let sealed_block = block.seal_slow();
-
-        Ok(Payload::new(attributes.payload_id(), sealed_block, U256::ZERO))
-    }
+        Attributes: PayloadBuilderAttributes;
 }
 
 /// Represents the outcome of committing withdrawals to the runtime database and post state.
